Added support for deseiralizaing Spyder v5.2+ script file format
In 5.2 the element base objects were updated to support content of types
still or source instead of maintaining only a source list.  I've updated
the SpyderClientLibrary to support this as well.  This shouldn't have
any breaking changes, but I did set some obsolete attributes on
properties specific to source or still elements that should now instead
reference the Content collections on elements.